backend/requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
psycopg2>=2.9
django-bootstrap-v5
djangorestframework>=3.13
djangorestframework-simplejwt
django-otp
django-formtools
django-two-factor-auth
phonenumbers
python-dotenv
django-oauth-toolkit
py-oauth2
os